tidy bsky pipeline tests, move helpers into dev-env
